NullPointerException in FileUtil.deleteDirectory() when deleting empty directories
git-svn-id: https://svn.code.sf.net/p/databenecommons/code/trunk@527 783448d1-dfc0-4521-9cb0-58c191f7d5bb

+++ src/main/java/org/databene/commons/FileUtil.java
@@ -141,11 +141,13 @@ public final class FileUtil {
 
     public static void deleteDirectory(File folder) {
     	File[] files = folder.listFiles();
-    	for (File file : files) {
-    		if (file.isDirectory())
-    			deleteDirectory(file);
-    		else
-    			file.delete();
+    	if (files != null) {
+	    	for (File file : files) {
+	    		if (file.isDirectory())
+	    			deleteDirectory(file);
+	    		else
+	    			file.delete();
+	    	}
     	}
     	folder.delete();
     }
